+++ src/claqgs.c
@@ -99,7 +99,7 @@ claqgs(SuperMatrix *A, float *r, float *c,
     /* Local variables */
     NCformat *Astore;
     complex   *Aval;
-    int i, j, irow;
+    long long i, j, irow;
     float large, small, cj;
     float temp;
